+++ src/main/java/de/dm/intellij/liferay/module/LiferayModuleListener.java
@@ -8,6 +8,7 @@ import com.intellij.psi.search.FilenameIndex;
 import com.intellij.psi.search.GlobalSearchScope;
 import de.dm.intellij.liferay.bnd.OsgiBndFileParser;
 import de.dm.intellij.liferay.hook.LiferayHookXmlParser;
+import de.dm.intellij.liferay.language.gulp.LiferayGulpfileParser;
 import de.dm.intellij.liferay.language.jsp.LiferayJspWebContentRootListener;
 import de.dm.intellij.liferay.theme.LiferayLookAndFeelXmlParser;
 import de.dm.intellij.liferay.theme.LiferayPackageJSONParser;
@@ -26,6 +27,7 @@ public class LiferayModuleListener implements ModuleListener {
             handleModuleFiles(project, module, "bnd.bnd");
             handleModuleFiles(project, module, "package.json");
             handleModuleFiles(project, module, "resources");
+            handleModuleFiles(project, module, "gulpfile.js");
         });
     }
 
@@ -47,6 +49,9 @@ public class LiferayModuleListener implements ModuleListener {
                 }
                 if (LiferayJspWebContentRootListener.isRelevantFile(virtualFile.getPath())) {
                     LiferayJspWebContentRootListener.handleChange(project, virtualFile);
+                }
+                if (LiferayGulpfileParser.isRelevantFile(virtualFile.getPath())) {
+                    LiferayGulpfileParser.handleChange(project, virtualFile);
                 }
             }
         }