+++ testdata/de/dm/intellij/liferay/language/gulp/LiferayGulpfileParserTest/gulpfile.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const gulp = require('gulp');
+const liferayThemeTasks = require('liferay-theme-tasks');
+
+liferayThemeTasks.registerTasks({
+    gulp: gulp,
+    pathSrc: './mysrc'
+});